--- a/.github/workflows/codeql-analysis.yml
+++ b/.github/workflows/codeql-analysis.yml
@@ -13,9 +13,7 @@ on:
       - "**/*.py"
       - "**/*.js"
       - "**/*.ts"
-      - requirements.txt
-      - requirements-dev.txt
-      - setup.cfg
+      - pyproject.toml
       - alembic.ini
 
   schedule: