--- a/.github/workflows/python-package.yaml
+++ b/.github/workflows/python-package.yaml
@@ -29,7 +29,7 @@ jobs:
         with:
           python-version: "3.9"
           cache: "pip"
-          cache-dependency-path: "requirements*.txt"
+          cache-dependency-path: "pyproject.toml"
 
       - uses: actions/setup-node@v4
         with: