--- a/.github/workflows/windows-tests.yaml
+++ b/.github/workflows/windows-tests.yaml
@@ -43,7 +43,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
-          cache-dependency-path: "requirements*.txt"
+          cache-dependency-path: "pyproject.toml"
 
       - name: Install packages
         run: |