--- a/pyproject.toml
+++ b/pyproject.toml
@@ -25,20 +25,33 @@ classifiers = [
 
 
 dependencies = [
+    "prefect[client]",
     "aiosqlite >= 0.17.0, < 1.0.0",
     "alembic >= 1.7.5, < 2.0.0",
-    "anyio >= 4.4.0, < 5.0.0",
     "apprise >= 1.1.0, < 2.0.0",
-    "asgi-lifespan >= 1.0, < 3.0",
     "asyncpg >= 0.23, < 1.0.0",
-    "cachetools >= 5.3, < 6.0",
     "click >= 8.0, < 8.2",
-    "cloudpickle >= 2.0, < 4.0",
-    "coolname >= 1.0.4, < 3.0.0",
-    "croniter >= 1.0.12, < 4.0.0",
     "cryptography >= 36.0.1",
     "dateparser >= 1.1.1, < 2.0.0",
     "docker >= 4.0, < 8.0",
+    "humanize >= 4.9.0, < 5.0.0",
+    "jinja2 >= 3.0.0, < 4.0.0",
+    "jinja2-humanize-extension >= 0.4.0",
+    "pytz >= 2021.1, < 2025",
+    "readchar >= 4.0.0, < 5.0.0",
+    "sqlalchemy[asyncio] >= 2.0, < 3.0.0",
+    "typer >= 0.12.0, != 0.12.2, < 0.13.0",
+]
+
+
+[project.optional-dependencies]
+client = [
+    "anyio >= 4.4.0, < 5.0.0",
+    "asgi-lifespan >= 1.0, < 3.0",
+    "cachetools >= 5.3, < 6.0",
+    "cloudpickle >= 2.0, < 4.0",
+    "coolname >= 1.0.4, < 3.0.0",
+    "croniter >= 1.0.12, < 4.0.0",
     "exceptiongroup >= 1.0.0",
     "fastapi >= 0.111.0, < 1.0.0",
     "fsspec >= 2022.5.0",
@@ -46,10 +59,7 @@ dependencies = [
     "griffe >= 0.49.0, <2.0.0",
     "httpcore >=1.0.5, < 2.0.0",
     "httpx[http2] >= 0.23, != 0.23.2",
-    "humanize >= 4.9.0, < 5.0.0",
     "importlib_metadata >= 4.4; python_version < '3.10'",
-    "jinja2 >= 3.0.0, < 4.0.0",
-    "jinja2-humanize-extension >= 0.4.0",
     "jsonpatch >= 1.32, < 2.0",
     "jsonschema >= 4.0.0, < 5.0.0",
     "orjson >= 3.7, < 4.0",
@@ -63,24 +73,17 @@ dependencies = [
     "pydantic_settings",
     "python_dateutil >= 2.8.2, < 3.0.0",
     "python-slugify >= 5.0, < 9.0",
-    "pytz >= 2021.1, < 2025",
     "pyyaml >= 5.4.1, < 7.0.0",
-    "readchar >= 4.0.0, < 5.0.0",
     "rfc3339-validator >= 0.1.4, < 0.2.0",
     "rich >= 11.0, < 14.0",
     "ruamel.yaml >= 0.17.0",
     "sniffio >=1.3.0, < 2.0.0",
-    "sqlalchemy[asyncio] >= 2.0, < 3.0.0",
     "toml >= 0.10.0",
-    "typer >= 0.12.0, != 0.12.2, < 0.13.0",
     "typing_extensions >= 4.5.0, < 5.0.0",
     "ujson >= 5.8.0, < 6.0.0",
     "uvicorn >=0.14.0, !=0.29.0",
     "websockets >= 10.4, < 14.0",
 ]
-
-
-[project.optional-dependencies]
 dev = [
     "ruff",
     "cairosvg",