--- a/src/prefect/cli/root.py
+++ b/src/prefect/cli/root.py
@@ -9,7 +9,6 @@ from importlib.metadata import PackageNotFoundError
 from importlib.metadata import version as import_version
 from typing import Any, Dict
 
-import pendulum
 import typer
 
 import prefect
@@ -111,10 +110,7 @@ async def version(
         "Version": prefect.__version__,
         "API version": SERVER_API_VERSION,
         "Python version": platform.python_version(),
-        "Git commit": prefect.__version_info__["full-revisionid"][:8],
-        "Built": pendulum.parse(
-            prefect.__version_info__["date"]
-        ).to_day_datetime_string(),
+        "Git commit": prefect.__version_info__[-1][:8],
         "OS/Arch": f"{sys.platform}/{platform.machine()}",
         "Profile": prefect.context.get_settings_context().profile.name,
     }