--- a/vermin.ini
+++ b/vermin.ini
@@ -1,0 +1,9 @@
+[vermin]
+make_paths_absolute = no
+backports = 
+    typing_extensions
+format = parsable
+eval_annotations = yes
+only_show_violations = yes
+targets = 3.9-
+exclusion_regex = ^src/prefect/_vendor/.*$|^src/prefect/utilities/compat\\.py$|^tests/workers/test_process_worker.py$|^tests/test_background_tasks.py$|^src/integrations/.*/tests/.*$