… requirements-client.txt
… requirements-dev.txt
… requirements.txt
… scripts/generate-lower-bounds.py
… setup.cfg
… setup.py
… versioneer.py