i started on this branch trying to improve import strategies but after profiling with `cProfile` i saw that a lot of time was being spent doing `versioneer` stuff and [read](https://github.com/pytroll/satpy/pull/856) that using `setuptools_scm` can improve this.

based on 
```
» pytest --benchmark-only --benchmark-min-rounds=10 benches/bench_import.py
```

| Test                     | Main (ms) | Branch (ms) |
|--------------------------|-----------|-------------|
| bench_import_prefect      | 66.04     | 0.66        |
| bench_import_prefect_flow | 314.68    | 242.98      |

I think there are many good reasons to switch to setuptools_scm over versioneer
- we already use it in collections
- versioneer is unmaintained
- edit all config in one place (instead of setup.py, setup.cfg and requirements*)